[fix/TK-193]: server에서 FormData 생성 에러 수정
