+++ apps/mobile/app/travel/[id]/manual/page.tsx
@@ -21,6 +21,33 @@ interface ManualRegisterSharedPaymentProps {
   };
 }
 
+const getPaymentDate = (date: string, time: string): string => {
+  const [amPm, hourMinute] = time.split(' ');
+
+  // 시간 형식이 잘못된 경우 처리
+  if (!hourMinute) {
+    throw new Error('Invalid time format');
+  }
+
+  const [hourStr, minuteStr] = hourMinute.split(':');
+
+  // 시간 또는 분이 없는 경우 처리
+  if (!hourStr || !minuteStr) {
+    throw new Error('Invalid time format');
+  }
+
+  let hour = parseInt(hourStr, 10);
+
+  // 오전/오후 처리
+  if (amPm === '오후' && hour !== 12) {
+    hour += 12; // 오후는 12를 더해줌
+  } else if (amPm === '오전' && hour === 12) {
+    hour = 0; // 오전 12시는 자정이므로 0으로 처리
+  }
+
+  return `${date} ${hour.toString().padStart(2, '0')}:${minuteStr.padStart(2, '0')}`;
+};
+
 export default function Page({ params }: ManualRegisterSharedPaymentProps) {
   const { id } = params;
   const router = useRouter();
@@ -87,8 +114,34 @@ export default function Page({ params }: ManualRegisterSharedPaymentProps) {
       return;
     }
 
+    const formDataToSend = new FormData();
+
+    // 필수값
+    formDataToSend.append(
+      'paymentDate',
+      getPaymentDate(formData.date, formData.time),
+    );
+    formDataToSend.append('storeName', formData.storeName);
+    formDataToSend.append('paymentAmount', formData.paymentAmount.toString());
+    formDataToSend.append('currencyUnit', formData.currencyUnit);
+    formDataToSend.append('isMainImage', formData.isMainImage.toString());
+    formDataToSend.append(
+      'paymentImage',
+      formData.paymentImage ? formData.paymentImage : new Blob(),
+    );
+    formDataToSend.append('paymentComment', formData.paymentComment);
+
+    if (formData.foreignPaymentAmount !== 0)
+      formDataToSend.append(
+        'foreignPaymentAmount',
+        formData.foreignPaymentAmount.toString(),
+      );
+
+    if (formData.exchangeRate !== 0)
+      formDataToSend.append('exchangeRate', formData.exchangeRate.toString());
+
     // 결제 내역 저장 요청
-    const response = await createManualSharedPayment(id, formData);
+    const response = await createManualSharedPayment(id, formDataToSend);
 
     if ('code' in response) {
       showToast.warning({
